Separate interceptors into files
